mod fill and linspace
